--- a/dynparquet/schema.go
+++ b/dynparquet/schema.go
@@ -1383,6 +1383,7 @@ func (c *dynColSet) Release() {
 }
 
 func (c *dynColSet) Merge(sets []map[string][]string) (o map[string][]string) {
+	//TODO:(gernest) use k-way merge
 	o = make(map[string][]string)
 	for i := range sets {
 		for k := range sets[i] {