This commit

- Adds test (There were no test covering this function)
- Introduce a new implementation for merging column sets that can be reused.
- Use `*sync.Pool` to reuse the new dynamic column set merger

Benchmark result

```
goos: darwin
goarch: amd64
pkg: github.com/polarsignals/frostdb/dynparquet
cpu: Intel(R) Core(TM) i7-7820HQ CPU @ 2.90GHz
                         │   old.txt   │               new.txt               │
                         │   sec/op    │   sec/op     vs base                │
MergeDynamicColumnSets-8   2.220µ ± 1%   1.330µ ± 1%  -40.09% (p=0.000 n=10)

                         │   old.txt   │              new.txt               │
                         │    B/op     │    B/op     vs base                │
MergeDynamicColumnSets-8   1104.0 ± 0%   544.0 ± 0%  -50.72% (p=0.000 n=10)

                         │   old.txt   │              new.txt               │
                         │  allocs/op  │ allocs/op   vs base                │
MergeDynamicColumnSets-8   16.000 ± 0%   6.000 ± 0%  -62.50% (p=0.000 n=10)
```