@brancz 2023-12-31
Dynamic columns are sorted intentionally, so we can do a k-way merge, we just hadn't implemented doing that yet.