@gernest 2023-12-31
@brancz 
> Dynamic columns are sorted intentionally, so we can do a k-way merge, we just hadn't implemented doing that yet.

Yes, I think this is described in commit message but didn't go into godoc for `MergeRecords `  #293

I know this is a statement, but I was wondering if you could provide some context relative to this PR.

**_note_** : in this PR behaviour is the same ,only implementation changed (dynamic columns are sorted as before)